requirements.txt
Flask==3.0.3
mysql-connector-python==9.1.0
python-dotenv==1.0.1
azure-ai-formrecognizer==3.3.3
azure-core==1.31.0
openai==1.52.0